requirements.txt
pandas
numpy
scikit-learn
torch
torchvision
gym
kafka-python
matplotlib
